+++ lib/src/model/field.dart
@@ -181,8 +181,4 @@ class Field extends ModelElement
 
   @override
   Inheritable? get overriddenElement => null;
-
-  @override
-  bool get hasHideConstantImplementation =>
-      definingEnclosingContainer.hasHideConstantImplementations;
 }
