+++ lib/src/model/getter_setter_combo.dart
@@ -58,12 +58,7 @@ mixin GetterSetterCombo on ModelElement {
   bool get isInherited;
 
   /// Whether this has a constant value which should be displayed.
-  bool get hasConstantValueForDisplay {
-    final element = this.element;
-    if (element is! ConstVariableElement) return false;
-    if (hasHideConstantImplementation) return false;
-    return element.constantInitializer != null;
-  }
+  bool get hasConstantValueForDisplay => false;
 
   Expression? get constantInitializer =>
       (element as ConstVariableElement).constantInitializer;
@@ -262,10 +257,6 @@ mixin GetterSetterCombo on ModelElement {
 
   // TODO(srawlins): This should be private.
   bool get writeOnly => hasPublicSetter && !hasPublicGetter;
-
-  /// True if the @hideConstantImplementations directive is present
-  /// in the defining enclosing element.
-  bool get hasHideConstantImplementation;
 
   @override
   late final Map<String, CommentReferable> referenceChildren = {
