+++ lib/src/model/top_level_variable.dart
@@ -82,8 +82,4 @@ class TopLevelVariable extends ModelElement
 
   @override
   Iterable<CommentReferable> get referenceParents => [definingLibrary];
-
-  @override
-  bool get hasHideConstantImplementation =>
-      definingLibrary.hasHideConstantImplementations;
 }
